requirements.txt
# requirements.txt

# Core dependencies
pywin32==308        # Windows COM interface for Word automation
pyyaml==6.0.1       # YAML configuration file handling
pytest==8.3.4       # Testing framework

# Development dependencies
black==24.2.0       # Code formatting
flake8==7.0.0       # Code linting
mypy==1.9.0         # Type checking

# Testing dependencies
pytest-cov==4.1.0   # Test coverage reporting
pytest-mock==3.12.0 # Mocking for tests

# Documentation
mkdocs==1.5.3       # Documentation generation
mkdocs-material==9.5.13  # Documentation theme

# Logging and debugging
logging==0.4.9.6    # Enhanced logging capabilities

# Type hints
typing-extensions==4.10.0  # Enhanced typing support

# Development tools
pip-tools==7.4.1    # Requirements file management